DocChat-RAG/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-classic
langchain-community
langchain-google-genai
langchain-huggingface
langchain-chroma
langchain-text-splitters
chromadb
pypdf
rank-bm25
sentence-transformers
